refactor: 특정 화면에 종속된 타입을 해당 VC파일 내부로 이동, NumberFormatterManager 구현, MovieRankMarkColor 수정
