--- BoxOffice/Controller/Enum/Section.swift
@@ -1,10 +1,0 @@
-//
-//  Section.swift
-//  BoxOffice
-//
-//  Created by 리지, kokkilE on 2023/04/05.
-//
-
-enum Section: Hashable {
-    case main
-}
