+++ BoxOffice/Controller/SelectDateViewController.swift
@@ -67,3 +67,14 @@ extension SelectDateViewController: UICalendarSelectionSingleDateDelegate {
         dismiss(animated: true)
     }
 }
+
+fileprivate enum AlertMessage {
+    case updating
+    
+    var description: String {
+        switch self {
+        case .updating:
+            return "오늘 날짜는 집계중입니다.\n다시 선택해주세요."
+        }
+    }
+}
